refactor(frontend): declutter by combining link components
